Nice pickup by FindBugs:
[M M IS] Inconsistent synchronization [IS2_INCONSISTENT_SYNC]

The fields of this class appear to be accessed inconsistently with respect to synchronization.  This bug report indicates that the bug pattern detector judged that 

The class contains a mix of locked and unlocked accesses, 
At least one locked access was performed by one of the class's own methods, and 
The number of unsynchronized field accesses (reads and writes) was no more than one third of all accesses, with writes being weighed twice as high as reads 
A typical bug matching this bug pattern is forgetting to synchronize one of the methods in a class that is intended to be thread-safe.